refactor: Change MobileOperator id to short
Changed MobileOperator id from int to short

+++ src/main/java/com/gtss/mnp_manager/models/MobileOperator.java
@@ -28,7 +28,7 @@ public class MobileOperator {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", updatable = false)
-    private int id;
+    private Short id;
 
     @NonNull
     @Column(name = "operator_name", nullable = false, length = 30)
